Move the token substitution logic to its own pkg
